PulsarAdministration accepts multiple customizers
See #432
